+++ tests/manualtest/run_operationtimeouts.py
@@ -79,7 +79,9 @@ class ClientTest(unittest.TestCase):
         # pylint: disable=global-variable-not-assigned,global-statement
         global args  # pylint: disable=invalid-name
 
-        self.host = args['host']  # pylint: disable=used-before-assignment
+        # pylint: disable=used-before-assignment
+        # pylint: disable=possibly-used-before-assignment
+        self.host = args['host']
         self.port = args['port']
         self.verbose = args['verbose']
         self.debug = args['debug']