+++ docs/changes.rst
@@ -38,7 +38,9 @@ Released: not yet
 
 **Bug fixes:**
 
-* Fixed safety issues up to 2024-05-13.
+* Fixed safety issues up to 2024-05-21.
+
+* Fixed new issues reported by Pylint 3.2.
 
 * Fix issue where pywbemlistener that fails startup can cause thread
   exception. See issue #3157